refactor(cli): finalize the custom inspector proxy for merging
